Started on #6 and #7
Added the "Create New Task"  fab button on the Day View. Clicking it
launches the NewTaskActivity. Having trouble sending the newly created
Task back to DayView (might need to make it Serializable?). Also might
need to change ArrayAdapter to hold an ArrayList, so that we can easily
add new Tasks to it.